pyStcratch/requirements.txt
# requirements.txt
# Web Server
flask>=3.0.0
gunicorn>=21.0.0

# FastAPI Stack
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Core dependencies - 使用更宽泛的版本号
scrapy>=2.11.0
aiohttp>=3.9.0
sqlalchemy>=2.0.23
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
requests>=2.31.0
fake-useragent>=1.4.0
loguru>=0.7.0
pandas>=2.0.0
twisted>=23.0.0
pyyaml>=6.0.0

# Database & Vector Store
supabase>=2.0.0
pinecone-client>=2.2.0

# AI/ML APIs
dashscope>=1.14.0
zhipuai>=2.1.5

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Playwright for browser automation (uncomment if needed)
# playwright==1.40.0

# Streamlit Frontend
streamlit>=1.28.0

# HTTP clients
httpx>=0.24.0

# Charts
plotly>=5.17.0
altair>=5.0.0

# Utilities
pydantic>=2.0.0
